+++ Src/Classes/SharedSettings.lua
@@ -17,6 +17,7 @@ sharedSettingsModule.defaults = {
 ---@field distance number
 
 ---@class BomSharedSettings Current character state snapshots per profile
+---@field DebugLogging boolean
 ---@field PlaySoundWhenTask string Play a sound when task list is not empty
 ---@field Minimap BomMinimapSettings
 ---@field SpellGreaterEqualThan table