+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## Buffomat Classic 2022.10.3
+## Buffomat Classic 2022.10.4, 10.5
 
 - Spell tab update is now hard limited in frequency to combat the raid lag.
 - Buff scan update is also hard limited in frequency for the same reason.
 - Horn of Winter rank 2 added
+- Option to play sound when Buffomat has something to do
+- Unit cache is less aggressive, this should help buffs to update faster
 
 ## Buffomat Classic 2022.10.1, -10.2